+++ Dockerfile
@@ -17,4 +17,5 @@ COPY . .
 
 EXPOSE 3000
 
-CMD [ "npm", "start" ]
+CMD [ "npm", "run", "start-backend" ]
+CMD ["npm", "run", "start-frontend"]
