+++ app.js
@@ -14,6 +14,15 @@ app.get('/', (req, res) =>{
     res.sendFile('index.html', {root: __dirname }); 
 });
 
+app.get('/results', (req, res) =>{
+    fetch('http://localhost:3500/results?number=' + req.query.number + '&street=' + req.query.street + '&type=' + req.query.type + '&city=' + req.query.city + '&state=' + req.query.state ) //'http://localhost:3500/results?number=2840&street=Tanglerose&type=Place&city=Eagle&state=ID' ) 
+         .then(res => {
+             return res.json(); // treat fetch response as a .json format and return to next promise
+         }).then( data => {
+             res.json(data); //console.log(data);
+         }).catch(error => console.log(error)); 
+ });
+
 // if endpoint not met (404 error), send message
 app.all('*', (req, res) => { 
     res.status(404).send('resource note found')
@@ -21,5 +30,5 @@ app.all('*', (req, res) => {
 
 // server now listening for attempts from a client to connect at port
 app.listen(port, () => { 
-    console.log(`Now listening in localhost`); 
+    console.log(`Frontend: Now listening`); 
 });
