+++ index.html
@@ -6,7 +6,7 @@
         <meta charset="urf-8">
         <meta name="viewport" content="width=device-width, initial-scale=1"> <!-- Webpage scalable for any device-->
         <title>Get Weather</title>
-        <script type="text/javascript" src="/fetch.js"></script>  
+        <!-- <script type="text/javascript" src="/fetch.js"></script>  -->
         <link rel="stylesheet" type="text/css" href="/styles.css">
     </head> 
     <body>
@@ -50,10 +50,9 @@
                 event.preventDefault();
                 const formData = new FormData(form); // get form data
                 let object = formDataToObject(formData); // store form data in object
-                let geoURL = "https://maps.googleapis.com/maps/api/geocode/json?address=";
-                let address = geoURL + object.number + "+" + object.street + "+" + object.type + ",+" + object.city + ",+" + object.state + "&key=";
-               
-                display(address); 
+                let geoURL = "https://maps.googleapis.com/maps/api/geocode/json?address="; 
+                let geoKeySign = "&key=";
+                display(object.number, object.street, object.type, object.city, object.state); 
             });
 
             // converts formData to object then returns
@@ -68,8 +67,14 @@
             }
 
             // display all weather ada after data() returns values
-            async function display(address){
-                    let value = await data(address);
+            async function display(number, street, type, city, state){
+                let value = await fetch('http://localhost:3000/results?number=' + number + '&street=' + street + '&type=' + type + '&city=' + city + '&state=' + state ) // 'http://localhost:3500/results?number=' + number + '&street=' + street + '&type=' + type + '&city='+ city + '&state=' + state ) 
+                        .then(res => {
+                            return res.json(); // treat fetch response as a .json format and return to next promise
+                        }).then( data => {
+                            return data;
+                            //console.log(data);
+                        }).catch(error => console.log(error)); 
                     document.getElementById("temp").innerHTML = value.data[0].coordinates[0].dates[0].value;
                     document.getElementById("max").innerHTML = value.data[4].coordinates[0].dates[0].value;
                     document.getElementById("min").innerHTML = value.data[6].coordinates[0].dates[0].value;
